+++ apps/pages/urls.py
@@ -6,6 +6,6 @@ urlpatterns = [
     path("", views.LandingPageView.as_view(), name="landing"),
     path("privacy-policy", views.PrivacyPolicyView.as_view(), name="privacy_policy"),
     path("terms-of-service", views.TermsOfServiceView.as_view(), name="terms_of_service"),
-    path("pricing", views.PricingView.as_view(), name="pricing"),
+    path("free-access", views.FreeAccessView.as_view(), name="free_access"),
     
 ]